refactor(front): move projects DTO types out of lib/api into types/api
Transport request/response DTO types and zod schemas are HTTP concerns and
do not belong next to business logic in lib/api. Move the `projects` domain
DTOs to `types/api/projects/*` (mirroring the lib/api module path), leaving
the business-logic functions and their internal helper types in place.

- metadata.ts, preferences.ts, tasks.ts: pure-contract files, moved wholesale.
- lookup.ts: removed re-export shim; consumers now import from types/api.
- list.ts, search.ts: DTOs extracted; functions stay. list.ts back-imports
  SearchProjectsResponseBody from types/api (correct lib -> types direction).
- Updated all consumers across front and front-api.

First batch of the lib/api DTO migration.
