+++ front-api/routes/w/[wId]/spaces/[spaceId]/project_metadata.ts
@@ -1,8 +1,4 @@
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agent";
-import type {
-  GetPodMetadataResponseBody,
-  PatchPodMetadataResponseBody,
-} from "@app/lib/api/projects/metadata";
 import { validatePinnedFramePath } from "@app/lib/api/projects/pinned_frame";
 import { getFeatureFlags } from "@app/lib/auth";
 import { ProjectMetadataResource } from "@app/lib/resources/project_metadata_resource";
@@ -12,6 +8,10 @@ import {
   stopProjectTodoWorkflow,
 } from "@app/temporal/project_task/client";
 import { PatchPodMetadataBodySchema } from "@app/types/api/internal/spaces";
+import type {
+  GetPodMetadataResponseBody,
+  PatchPodMetadataResponseBody,
+} from "@app/types/api/projects/metadata";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
