+++ front-api/routes/w/[wId]/spaces/[spaceId]/star.ts
@@ -1,6 +1,6 @@
-import type { PostUserPodStarResponseBody } from "@app/lib/api/projects/preferences";
-import { PostUserPodStarBodySchema } from "@app/lib/api/projects/preferences";
 import { UserProjectPreferencesResource } from "@app/lib/resources/user_project_preferences_resource";
+import type { PostUserPodStarResponseBody } from "@app/types/api/projects/preferences";
+import { PostUserPodStarBodySchema } from "@app/types/api/projects/preferences";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
