+++ front-api/routes/w/[wId]/spaces/projects-lookup.ts
@@ -1,8 +1,6 @@
-import {
-  enrichProjectsWithMetadata,
-  type SpacesLookupResponseBody,
-} from "@app/lib/api/projects/list";
+import { enrichProjectsWithMetadata } from "@app/lib/api/projects/list";
 import { SpaceResource } from "@app/lib/resources/space_resource";
+import type { SpacesLookupResponseBody } from "@app/types/api/projects/list";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
