--- front/lib/api/projects/lookup.ts
@@ -1,4 +1,0 @@
-export type {
-  SearchProjectsResponseBody,
-  SpacesLookupResponseBody,
-} from "@app/lib/api/projects/list";
